breaking: NetworkWriter inherits from Stream for compatibility with third party libraries. => .Position is now long, so users may need to cast to (int) here and there.

+++ Assets/Mirror/Runtime/NetworkConnection.cs
@@ -131,7 +131,7 @@ namespace Mirror
             {
                 // pack message and send allocation free
                 MessagePacking.Pack(message, writer);
-                NetworkDiagnostics.OnSend(message, channelId, writer.Position, 1);
+                NetworkDiagnostics.OnSend(message, channelId, (int)writer.Position, 1);
                 Send(writer.ToArraySegment(), channelId);
             }
         }
